platform/core/src/main/java/com/anthem/platform/core/models/DeepLinks.java
package com.anthem.platform.core.models;

import javax.inject.Inject;

import org.apache.sling.api.resource.Resource;
import org.apache.sling.models.annotations.Model;
import org.apache.sling.models.annotations.Optional;

@Model(adaptables = Resource.class)
public class DeepLinks {

	@Inject
	@Optional
	private String paramName;

	@Inject
	@Optional
	private String paramValue;

	/**
	 * @return the paramName
	 */
	public String getParamName() {
		return paramName;
	}

	/**
	 * @return the paramValue
	 */
	public String getParamValue() {
		return paramValue;
	}

}